// Frontend/quran-dashboard-ui/src/app/core/api/generated/models/linking-prepared-detail-page-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { LinkingPreparedDetailItemDto } from '../models/linking-prepared-detail-item-dto';
export interface LinkingPreparedDetailPageDto {
  detailKind: string;
  filter: string;
  items: Array<LinkingPreparedDetailItemDto>;
  linkingDataRevision: number;
  page: number;
  pageSize: number;
  preflightId: string;
  preparedSourceId: number | null;
  totalItems: number;
  totalPages: number;
}