// Frontend/quran-dashboard-ui/src/app/core/api/generated/models/linking-prepared-detail-page-dto-api-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { LinkingPreparedDetailPageDto } from '../models/linking-prepared-detail-page-dto';
export interface LinkingPreparedDetailPageDtoApiResponse {
  data: LinkingPreparedDetailPageDto | null;
  errors: Array<string> | null;
  isSuccess: boolean;
  message: string | null;
}