// Frontend/quran-dashboard-ui/src/app/core/api/generated/models/linking-prepared-preflight-status-dto-api-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { LinkingPreparedPreflightStatusDto } from '../models/linking-prepared-preflight-status-dto';
export interface LinkingPreparedPreflightStatusDtoApiResponse {
  data: LinkingPreparedPreflightStatusDto | null;
  errors: Array<string> | null;
  isSuccess: boolean;
  message: string | null;
}